apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-multipart==0.0.20
python-jose[cryptography]==3.4.0
passlib==1.7.4
python-dotenv==1.1.1
pyarrow>=14.0.0
requests>=2.31.0
yfinance>=0.2.0
py_vollib>=1.0.1
pandas>=2.0.0
vaderSentiment>=3.3.2
# pytrends>=5.0.0  # Optional: Google Trends. Uncomment for search volume in sentiment.
# transformers>=4.30.0  # Optional: FinBERT. Uncomment for ensemble NLP.
# torch>=2.0.0  # Optional: Required for FinBERT.